// src/store/actions/cities/interfaces.ts
export enum getCitiesType {
  GET_CITIES = 'GET_CITIES',
  GET_CITIES_FAIL = 'GET_CITIES_FAIL',
  GET_CITIES_SUCCESS = 'GET_CITIES_SUCCESS',
  SHOW_WARNING = 'SHOW_WARNING',
  CANCEL_DELETE = 'CANCEL_DELETE',
  ADD_CITY_FAIL = 'ADD_CITY_FAIL',
  RESET_CITIES_ERRORS = 'RESET_CITIES_ERRORS',
  SHOW_CITY = 'SHOW_CITY',
  RESET_EDITING = 'RESET_EDITING',
  EDIT_CITY_FAIL = 'EDIT_CITY_FAIL',
}